refactor: improve NFT View page on desktop
