+++ apps/next/src/pages/_app.tsx
@@ -92,9 +92,9 @@ export default function App({ Component, pageProps, router }: AppProps) {
             }
           />
           <View
-            tw="min-h-screen items-center"
-            // @ts-ignore currently overflow-x-hidden doesn't support web
-            style={{ overflowX: "hidden" }}
+            tw="items-center"
+            // @ts-ignore currently overflow-x-hidden doesn't support web, and minHeight we need to subtract Header height.
+            style={{ overflowX: "hidden", minHeight: "calc(100vh - 64px)" }}
           >
             <Component {...pageProps} />
           </View>
