+++ packages/app/components/swipe-list.tsx
@@ -1,66 +1,34 @@
-import { memo, useCallback, useMemo, useRef, Suspense, useState } from "react";
+import { useCallback, useMemo, useRef } from "react";
 import {
   Dimensions,
   FlatList,
   Platform,
-  Pressable,
   RefreshControl,
   useWindowDimensions,
 } from "react-native";
 
-import { BlurView } from "expo-blur";
-import Reanimated, {
+import {
   useAnimatedStyle,
   useSharedValue,
   withTiming,
 } from "react-native-reanimated";
 
-import {
-  useBlurredBackgroundStyles,
-  useIsDarkMode,
-} from "@showtime-xyz/universal.hooks";
-import { Share } from "@showtime-xyz/universal.icon";
-import { Image } from "@showtime-xyz/universal.image";
 import { useSafeAreaFrame } from "@showtime-xyz/universal.safe-area";
-import { Skeleton } from "@showtime-xyz/universal.skeleton";
 import { tw } from "@showtime-xyz/universal.tailwind";
-import { Text } from "@showtime-xyz/universal.text";
-import { View } from "@showtime-xyz/universal.view";
 
-import { Description } from "app/components/card/rows/description";
-import { Creator } from "app/components/card/rows/elements/creator";
-import { Owner } from "app/components/card/rows/owner";
-import { Social } from "app/components/card/social";
-import { ClaimButton } from "app/components/claim/claim-button";
-import { GiftButton } from "app/components/claim/gift-button";
-import { CommentButton } from "app/components/feed/comment-button";
-import { FeedItemTapGesture } from "app/components/feed/feed-item-tap-gesture";
-import { Like } from "app/components/feed/like";
-import { LikedBy } from "app/components/liked-by";
-import { Media } from "app/components/media";
-import { NFTDropdown } from "app/components/nft-dropdown";
+import { FeedItem } from "app/components/feed-item";
 import { MAX_HEADER_WIDTH } from "app/constants/layout";
-import { LikeContextProvider } from "app/context/like-context";
 import { VideoConfigContext } from "app/context/video-config-context";
-import {
-  CreatorEditionResponse,
-  useCreatorCollectionDetail,
-} from "app/hooks/use-creator-collection-detail";
 import { useIsMobileWeb } from "app/hooks/use-is-mobile-web";
-import { useShareNFT } from "app/hooks/use-share-nft";
 import { useUser } from "app/hooks/use-user";
-import { Blurhash } from "app/lib/blurhash";
 import { useHeaderHeight } from "app/lib/react-navigation/elements";
 import { useNavigation, useScrollToTop } from "app/lib/react-navigation/native";
 import { DataProvider, LayoutProvider } from "app/lib/recyclerlistview";
 import type { NFT } from "app/types";
-import { getMediaUrl } from "app/utilities";
 
 import { ViewabilityTrackerRecyclerList } from "./viewability-tracker-swipe-list";
 
 const { height: screenHeight, width: screenWidth } = Dimensions.get("screen");
-const mediaMaxHeightRelativeToScreen = 1;
-const NFT_DETAIL_WIDTH = 380;
 const SCROLL_BAR_WIDTH = 15;
 const MOBILE_WEB_TABS_HEIGHT = 50;
 const MOBILE_WEB_BOTTOM_NAV_HEIGHT = 64;
@@ -78,7 +46,7 @@ type Props = {
 export const SwipeList = ({
   data,
   fetchMore,
-  isRefreshing,
+  isRefreshing = false,
   refresh,
   initialScrollIndex = 0,
   bottomPadding = 0,
@@ -254,274 +222,5 @@ export const SwipeList = ({
         layoutSize={layoutSize}
       />
     </VideoConfigContext.Provider>
-  );
-};
-
-export const FeedItem = memo(
-  ({
-    nft,
-    bottomPadding = 0,
-    bottomMargin = 0,
-    itemHeight,
-    hideHeader,
-    showHeader,
-    toggleHeader,
-    detailStyle,
-    listId,
-  }: {
-    nft: NFT;
-    detailStyle: any;
-    showHeader: any;
-    hideHeader: any;
-    toggleHeader: any;
-    bottomPadding: number;
-    bottomMargin?: number;
-    itemHeight: number;
-    listId?: number;
-  }) => {
-    const [detailHeight, setDetailHeight] = useState(0);
-    const { width: windowWidth } = useWindowDimensions();
-    const { data: edition } = useCreatorCollectionDetail(
-      nft.creator_airdrop_edition_address
-    );
-    const blurredBackgroundStyles = useBlurredBackgroundStyles(95);
-
-    const isCreatorDrop = !!nft.creator_airdrop_edition_address;
-
-    const feedItemStyle = {
-      height: itemHeight,
-      width: windowWidth,
-    };
-
-    let mediaHeight =
-      windowWidth /
-      (isNaN(Number(nft.token_aspect_ratio))
-        ? 1
-        : Number(nft.token_aspect_ratio));
-
-    const mediaContainerHeight = Math.min(
-      mediaHeight,
-      feedItemStyle.height * mediaMaxHeightRelativeToScreen
-    );
-
-    mediaHeight = Math.min(mediaHeight, mediaContainerHeight);
-
-    const isDark = useIsDarkMode();
-    const tint = isDark ? "dark" : "light";
-    const mediaWidth = useMemo(() => {
-      if (windowWidth >= MAX_HEADER_WIDTH) {
-        return MAX_HEADER_WIDTH - NFT_DETAIL_WIDTH;
-      }
-
-      return windowWidth - NFT_DETAIL_WIDTH;
-    }, [windowWidth]);
-
-    if (windowWidth >= 768) {
-      return (
-        <LikeContextProvider nft={nft} key={nft.nft_id}>
-          <View tw="h-full w-full max-w-screen-2xl flex-row">
-            <View
-              style={[
-                tw.style(
-                  "flex-1 items-center justify-center bg-gray-100 dark:bg-black"
-                ),
-                {
-                  width: 100,
-                },
-              ]}
-            >
-              <Media
-                item={nft}
-                numColumns={1}
-                tw={`h-[${mediaHeight}px] w-[${mediaWidth}px]`}
-                resizeMode="contain"
-              />
-            </View>
-            <View
-              style={[
-                tw.style(
-                  "bg-white dark:bg-black shadow-lg shadow-black/5 dark:shadow-white/50"
-                ),
-                {
-                  width: NFT_DETAIL_WIDTH,
-                },
-              ]}
-            >
-              <Social nft={nft} />
-              <LikedBy nft={nft} />
-              <View tw="my-4 mr-4 flex-row justify-between px-4">
-                <Text tw="font-space-bold text-lg text-black dark:text-white md:text-2xl">
-                  {nft.token_name}
-                </Text>
-                <Suspense fallback={<Skeleton width={24} height={24} />}>
-                  <NFTDropdown nft={nft} listId={listId} />
-                </Suspense>
-              </View>
-              <Description nft={nft} />
-              <View tw="flex-row items-center justify-between px-4">
-                <Creator nft={nft} />
-                <Owner nft={nft} price={false} />
-              </View>
-              <View tw="px-4 py-4">
-                {isCreatorDrop && edition ? (
-                  <ClaimButton edition={edition} />
-                ) : null}
-                {/* {!isCreatorDrop ? <BuyButton nft={nft} /> : null} */}
-              </View>
-
-              {/* Comments */}
-            </View>
-          </View>
-        </LikeContextProvider>
-      );
-    }
-
-    return (
-      <LikeContextProvider nft={nft} key={nft.nft_id}>
-        <View tw="w-full flex-1">
-          {Platform.OS !== "web" && (
-            <View>
-              {nft.blurhash ? (
-                <Blurhash
-                  blurhash={nft.blurhash}
-                  decodeWidth={16}
-                  decodeHeight={16}
-                  decodeAsync={true}
-                  style={tw.style("w-full h-full")}
-                />
-              ) : (
-                <Image
-                  tw="h-full w-full"
-                  source={{
-                    uri: getMediaUrl({ nft, stillPreview: true }),
-                  }}
-                />
-              )}
-            </View>
-          )}
-
-          <FeedItemTapGesture
-            toggleHeader={toggleHeader}
-            showHeader={showHeader}
-          >
-            <View
-              tw={`absolute justify-center`}
-              style={{
-                height: Platform.select({
-                  web: itemHeight - bottomPadding - detailHeight,
-                  default: itemHeight - bottomPadding - 50,
-                }),
-              }}
-            >
-              <Media
-                item={nft}
-                numColumns={1}
-                tw={`h-[${mediaHeight}px] w-[${windowWidth}px]`}
-                resizeMode="contain"
-                onPinchStart={hideHeader}
-                onPinchEnd={showHeader}
-              />
-            </View>
-          </FeedItemTapGesture>
-
-          <Reanimated.View
-            style={[
-              tw.style("z-1 absolute right-0 left-0"),
-              detailStyle,
-              { bottom: bottomMargin },
-            ]}
-            onLayout={({
-              nativeEvent: {
-                layout: { height },
-              },
-            }) => {
-              if (Platform.OS !== "web") return;
-              setDetailHeight(height);
-            }}
-          >
-            <BlurView
-              tint={tint}
-              intensity={100}
-              style={{
-                // @ts-ignore
-                ...blurredBackgroundStyles,
-                ...tw.style(
-                  "bg-white bg-opacity-20 dark:bg-black dark:bg-opacity-20"
-                ),
-              }}
-            >
-              <NFTDetails edition={edition} nft={nft} listId={listId} />
-              <View
-                tw={`${
-                  bottomPadding && bottomPadding !== 0
-                    ? `h-[${bottomPadding - 1}px]`
-                    : "h-0"
-                }`}
-              />
-            </BlurView>
-          </Reanimated.View>
-        </View>
-      </LikeContextProvider>
-    );
-  }
-);
-FeedItem.displayName = "FeedItem";
-
-const NFTDetails = ({
-  nft,
-  listId,
-  edition,
-}: {
-  nft: NFT;
-  edition?: CreatorEditionResponse;
-  listId?: number;
-}) => {
-  const { shareNFT } = useShareNFT();
-  const isCreatorDrop = !!nft.creator_airdrop_edition_address;
-
-  return (
-    <View>
-      <View tw="flex-row items-center justify-between px-4">
-        <Creator nft={nft} shouldShowCreatorIndicator={false} />
-        {isCreatorDrop && edition ? <ClaimButton edition={edition} /> : null}
-        {/* {!isCreatorDrop ? <BuyButton nft={nft} /> : null} */}
-      </View>
-
-      <View tw="px-4">
-        <Text tw="font-space-bold text-lg dark:text-white" numberOfLines={3}>
-          {nft.token_name}
-        </Text>
-
-        <View tw="h-4" />
-
-        <View tw="flex-row justify-between">
-          <View tw="flex-row">
-            <Like nft={nft} />
-            <View tw="w-6" />
-            <CommentButton nft={nft} />
-            <View tw="w-6" />
-            <GiftButton nft={nft} />
-          </View>
-
-          <View tw="flex-row">
-            <Pressable onPress={() => shareNFT(nft)}>
-              <Share
-                height={22}
-                width={22}
-                // @ts-ignore
-                color={tw.style("bg-gray-900 dark:bg-white").backgroundColor}
-              />
-            </Pressable>
-            <View tw="w-8" />
-
-            <Suspense fallback={<Skeleton width={24} height={24} />}>
-              <NFTDropdown nft={nft} listId={listId} />
-            </Suspense>
-          </View>
-        </View>
-      </View>
-
-      <View tw="h-4" />
-    </View>
   );
 };
